+++ src/components/SelectBox/SelectBox.tsx
@@ -6,17 +6,7 @@
 
 import { Check, ChevronDown, Search } from 'lucide-react'
 import { cva, type VariantProps } from 'class-variance-authority'
-import { useSelect } from './hooks/useSelect'
-
-
-// export type SelectBoxProps = {
-//   options: {
-//     lists: SelectBoxOptions[]
-//     search?: boolean
-//   }
-//   value: SelectBoxOptions | null
-//   onChange: (option: SelectBoxOptions|null) => void
-// } & VariantProps<typeof wrapperVariants>
+import { useSelectBox } from './hooks/useSelectBox'
 
 const wrapperVariants = cva('relative h-12', {
   variants: {
@@ -32,33 +22,34 @@ const wrapperVariants = cva('relative h-12', {
   },
 })
 export type SelectBoxProps = {
-  selectBoxHook: ReturnType<typeof useSelect>
+  selectBoxHook: ReturnType<typeof useSelectBox>
 } & VariantProps<typeof wrapperVariants>
 
 export function SelectBox({ size, selectBoxHook }: SelectBoxProps) {
   const {
-    selectedValue,
+    selectedOption,
     isOpen,
     searchValue,
-    options,
+    filteredOptions,
+    selectedText,
     search,
-    handleSelect,
+    containerRef,
+    handleClickOption,
     handleToggle,
-    handleSearch
+    handleSearch,
   } = selectBoxHook
 
   return (
-    <div className={wrapperVariants({ size })}>
+    <div className={wrapperVariants({ size })} ref={containerRef}>
       <div
-        className="text-primary-100 border-primary-100 text-yds-b1 absolute top-0 left-0 flex h-full w-full cursor-pointer items-center justify-between rounded-lg border-2 p-3"
+        className="text-primary-100 border-primary-100 text-yds-b1 flex h-full w-full cursor-pointer items-center justify-between rounded-lg border-2 p-3"
         onClick={handleToggle}
       >
-        {!selectedValue && '선택'}
-        {selectedValue && selectedValue.label}
+        {selectedText}
         <ChevronDown className="text-primary-100 transition-transform duration-300" />
       </div>
       {isOpen && (
-        <div className="bg-background-secondary border-primary-100 absolute top-14 left-0 flex h-auto w-full flex-col gap-2 rounded-lg border-2 p-3">
+        <div className="bg-background-secondary border-primary-100 z-select-box-dropdown absolute top-14 left-0 flex w-full flex-col gap-2 rounded-lg border-2 p-3">
           {search && (
             <div className="border-primary-100 flex h-[40px] items-center border-y-2">
               <Search className="text-primary-100" size={20} />
@@ -71,18 +62,18 @@ export function SelectBox({ size, selectBoxHook }: SelectBoxProps) {
               />
             </div>
           )}
-          {options
-            .filter(option => option.label.includes(searchValue))
-            .map(option => (
+          <div className="scrollbar-hide flex h-auto max-h-[200px] flex-col gap-2 overflow-y-scroll">
+            {filteredOptions.map(option => (
               <div
                 key={option.label}
                 className="text-yds-c1m hover:bg-background-primary flex cursor-pointer items-center justify-between text-white"
-                onClick={() => handleSelect(option)}
+                onClick={() => handleClickOption(option)}
               >
-                {option.label}
-                {selectedValue?.value === option.value && <Check className="text-primary-100" size={20} />}
+                {option.value}
+                {selectedOption.value === option.value && <Check className="text-primary-100" size={20} />}
               </div>
             ))}
+          </div>
         </div>
       )}
     </div>